requirements.txt
numpy
pandas
matplotlib
opencv-python-headless
scikit-learn
scikit-image
torch
torchvision
streamlit
pillow
seaborn
kagglehub

